fix remove

+++ jetgraphs/JetGraphDataset.py
@@ -95,13 +95,13 @@ class JetGraphDatasetInMemory_v2(InMemoryDataset):
         
         # Remove anything which is not Signal_v6, Background2_v6, Background3_v6 or download from raw_dir
         possibly_trash = [os.path.join(self.raw_dir,subdir) for subdir in os.listdir(self.raw_dir)]
-        trash = [x for x in possibly_trash if not "v6" in x and not "download" in x]
-        print("Removing: ", trash)
-        for t in trash:
-            try:
-                os.remove(t)
-            except:
-                os.rmdir(t)
+        print("Removing: ", possibly_trash)
+        for t in possibly_trash:
+            if "v6" not in t and "download" not in t:
+                try:
+                    os.remove(t)
+                except:
+                    os.rmdir(t)
 
         print("Renaming files...")
         # Add the 'a0' prefix to files for layer 0.
